fix(inspect): format the chart time-axis by query step, shared with dashboards
The inspect board charts showed raw epoch-ms on the x-axis and tooltip header. Format each bucket by the query step the same way the layer-dashboard line widgets do (DAY → MM-DD, HOUR → MM-DD HH:00, MINUTE → HH:MM, browser-local). The per-step formatter is extracted to a shared `bucketTimeLabel` in utils/formatters and used by both the dashboard widgets and the inspect chart, so they can't drift.
